Added CFDictionary because I was curious. It won't make my published benchmarks.
It has a bad problem for benchmarks because I'm creating two dictionaries with a pre-size (because it supports it). This will eat twice as much memory as it should in the benchmarks.

A casual spot test run for the get good test with 30000000 (my system was doing other stuff so not a clean number), has the string test at 41.410724 (slightly slower than Perl speed) and the int test at 8.152435 (which beats STL unordered map). For the string test, like all the other tests, the string impedance mismatch rears its ugly head again, this time in the conversion from const char* to CFStringRef.